refactor: remove commands directory and move git into core
- Removed src/commands/ directory as all files were just re-exporting from core
- Moved version.ts to src/core/ since it contains actual implementation
- Moved src/git/ to src/core/git/ to consolidate all business logic under core
- Updated all imports to reflect new paths
- Updated CLAUDE.md to reflect the simplified structure

This further simplifies the architecture by eliminating unnecessary indirection
and consolidating all business logic under the core directory.
