--- src/commands/create.ts
@@ -1,1 +1,0 @@
-export { createWorktree } from "../core/worktree/create.ts";
