--- src/commands/delete.ts
@@ -1,1 +1,0 @@
-export { deleteWorktree } from "../core/worktree/delete.ts";
