--- src/commands/exec.ts
@@ -1,1 +1,0 @@
-export { execInWorktree } from "../core/process/exec.ts";
