--- src/commands/list.test.ts
@@ -1,271 +1,0 @@
-import { deepStrictEqual, strictEqual } from "node:assert";
-import { before, describe, it, mock } from "node:test";
-
-describe("listWorktrees", () => {
-  let accessMock: ReturnType<typeof mock.fn>;
-  let readdirMock: ReturnType<typeof mock.fn>;
-  let execMock: ReturnType<typeof mock.fn>;
-  let listWorktrees: typeof import("../core/worktree/list.ts").listWorktrees;
-
-  before(async () => {
-    accessMock = mock.fn();
-    readdirMock = mock.fn();
-    execMock = mock.fn();
-
-    mock.module("node:fs/promises", {
-      namedExports: {
-        access: accessMock,
-        readdir: readdirMock,
-      },
-    });
-
-    mock.module("node:child_process", {
-      namedExports: {
-        exec: execMock,
-      },
-    });
-
-    mock.module("node:util", {
-      namedExports: {
-        promisify: (fn: unknown) => fn,
-      },
-    });
-
-    mock.module("../core/worktree/validate.ts", {
-      namedExports: {
-        validatePhantomDirectoryExists: mock.fn((gitRoot: string) => {
-          if (gitRoot === "/test/repo") {
-            return Promise.resolve(
-              accessMock(`${gitRoot}/.git/phantom/worktrees`)
-                .then(() => true)
-                .catch(() => false),
-            );
-          }
-          return Promise.resolve(false);
-        }),
-        listValidWorktrees: mock.fn((gitRoot: string) => {
-          return readdirMock(`${gitRoot}/.git/phantom/worktrees`)
-            .then((files: string[]) => files)
-            .catch(() => []);
-        }),
-      },
-    });
-
-    mock.module("../core/paths.ts", {
-      namedExports: {
-        getWorktreePath: mock.fn(
-          (gitRoot: string, name: string) =>
-            `${gitRoot}/.git/phantom/worktrees/${name}`,
-        ),
-      },
-    });
-
-    ({ listWorktrees } = await import("../core/worktree/list.ts"));
-  });
-
-  it("should return empty array when worktrees directory doesn't exist", async () => {
-    accessMock.mock.resetCalls();
-    readdirMock.mock.resetCalls();
-    execMock.mock.resetCalls();
-
-    // Mock getGitRoot
-    execMock.mock.mockImplementation((cmd: string) => {
-      if (cmd === "git rev-parse --show-toplevel") {
-        return Promise.resolve({ stdout: "/test/repo\n", stderr: "" });
-      }
-      return Promise.resolve({ stdout: "", stderr: "" });
-    });
-
-    // Mock worktrees directory doesn't exist
-    accessMock.mock.mockImplementation((path: string) => {
-      if (path === "/test/repo/.git/phantom/worktrees") {
-        return Promise.reject(new Error("ENOENT"));
-      }
-      return Promise.resolve();
-    });
-
-    const result = await listWorktrees("/test/repo");
-
-    strictEqual(result.success, true);
-    deepStrictEqual(result.worktrees, []);
-    strictEqual(
-      result.message,
-      "No worktrees found (worktrees directory doesn't exist)",
-    );
-  });
-
-  it("should return empty array when worktrees directory is empty", async () => {
-    accessMock.mock.resetCalls();
-    readdirMock.mock.resetCalls();
-    execMock.mock.resetCalls();
-
-    // Mock getGitRoot
-    execMock.mock.mockImplementation((cmd: string) => {
-      if (cmd === "git rev-parse --show-toplevel") {
-        return Promise.resolve({ stdout: "/test/repo\n", stderr: "" });
-      }
-      return Promise.resolve({ stdout: "", stderr: "" });
-    });
-
-    // Mock worktrees directory exists but is empty
-    accessMock.mock.mockImplementation(() => Promise.resolve());
-    readdirMock.mock.mockImplementation(() => Promise.resolve([]));
-
-    const result = await listWorktrees("/test/repo");
-
-    strictEqual(result.success, true);
-    deepStrictEqual(result.worktrees, []);
-    strictEqual(result.message, "No worktrees found");
-  });
-
-  it("should list worktrees with clean status", async () => {
-    accessMock.mock.resetCalls();
-    readdirMock.mock.resetCalls();
-    execMock.mock.resetCalls();
-
-    // Mock getGitRoot and git commands
-    execMock.mock.mockImplementation(
-      (cmd: string, options?: { cwd?: string }) => {
-        if (cmd === "git rev-parse --show-toplevel") {
-          return Promise.resolve({ stdout: "/test/repo\n", stderr: "" });
-        }
-        if (cmd.includes("git -C") && cmd.includes("branch --show-current")) {
-          if (cmd.includes("test-worktree-1")) {
-            return Promise.resolve({ stdout: "feature/test\n", stderr: "" });
-          }
-          if (cmd.includes("test-worktree-2")) {
-            return Promise.resolve({ stdout: "main\n", stderr: "" });
-          }
-        }
-        if (cmd.includes("git -C") && cmd.includes("status --porcelain")) {
-          return Promise.resolve({ stdout: "", stderr: "" }); // Clean status
-        }
-        return Promise.resolve({ stdout: "", stderr: "" });
-      },
-    );
-
-    // Mock worktrees directory and contents
-    accessMock.mock.mockImplementation(() => Promise.resolve());
-    readdirMock.mock.mockImplementation(() =>
-      Promise.resolve(["test-worktree-1", "test-worktree-2"]),
-    );
-
-    const result = await listWorktrees("/test/repo");
-
-    strictEqual(result.success, true);
-    strictEqual(result.worktrees?.length, 2);
-    strictEqual(result.worktrees?.[0].name, "test-worktree-1");
-    strictEqual(result.worktrees?.[0].branch, "feature/test");
-    strictEqual(result.worktrees?.[0].isClean, true);
-    strictEqual(result.worktrees?.[1].name, "test-worktree-2");
-    strictEqual(result.worktrees?.[1].branch, "main");
-    strictEqual(result.worktrees?.[1].isClean, true);
-  });
-
-  it("should list worktrees with dirty status", async () => {
-    accessMock.mock.resetCalls();
-    readdirMock.mock.resetCalls();
-    execMock.mock.resetCalls();
-
-    // Mock getGitRoot and git commands
-    execMock.mock.mockImplementation(
-      (cmd: string, options?: { cwd?: string }) => {
-        if (cmd === "git rev-parse --show-toplevel") {
-          return Promise.resolve({ stdout: "/test/repo\n", stderr: "" });
-        }
-        if (cmd.includes("git -C") && cmd.includes("branch --show-current")) {
-          return Promise.resolve({ stdout: "feature/dirty\n", stderr: "" });
-        }
-        if (cmd.includes("git -C") && cmd.includes("status --porcelain")) {
-          return Promise.resolve({
-            stdout: " M file1.ts\n?? file2.ts\n",
-            stderr: "",
-          }); // Dirty status with 2 files
-        }
-        return Promise.resolve({ stdout: "", stderr: "" });
-      },
-    );
-
-    // Mock worktrees directory and contents
-    accessMock.mock.mockImplementation(() => Promise.resolve());
-    readdirMock.mock.mockImplementation(() =>
-      Promise.resolve(["dirty-phantom"]),
-    );
-
-    const result = await listWorktrees("/test/repo");
-
-    strictEqual(result.success, true);
-    strictEqual(result.worktrees?.length, 1);
-    strictEqual(result.worktrees?.[0].name, "dirty-phantom");
-    strictEqual(result.worktrees?.[0].branch, "feature/dirty");
-    strictEqual(result.worktrees?.[0].isClean, false);
-  });
-
-  it("should handle git command errors gracefully", async () => {
-    accessMock.mock.resetCalls();
-    readdirMock.mock.resetCalls();
-    execMock.mock.resetCalls();
-
-    // Mock getGitRoot and failing git commands
-    execMock.mock.mockImplementation((cmd: string) => {
-      if (cmd === "git rev-parse --show-toplevel") {
-        return Promise.resolve({ stdout: "/test/repo\n", stderr: "" });
-      }
-      if (cmd.includes("git -C") && cmd.includes("branch --show-current")) {
-        return Promise.reject(new Error("Not a git repository"));
-      }
-      if (cmd.includes("git -C") && cmd.includes("status --porcelain")) {
-        return Promise.reject(new Error("Git command failed"));
-      }
-      return Promise.resolve({ stdout: "", stderr: "" });
-    });
-
-    // Mock worktrees directory and contents
-    accessMock.mock.mockImplementation(() => Promise.resolve());
-    readdirMock.mock.mockImplementation(() =>
-      Promise.resolve(["error-phantom"]),
-    );
-
-    const result = await listWorktrees("/test/repo");
-
-    strictEqual(result.success, true);
-    strictEqual(result.worktrees?.length, 1);
-    strictEqual(result.worktrees?.[0].name, "error-phantom");
-    strictEqual(result.worktrees?.[0].branch, "unknown");
-    strictEqual(result.worktrees?.[0].isClean, true);
-  });
-
-  it("should handle detached HEAD state", async () => {
-    accessMock.mock.resetCalls();
-    readdirMock.mock.resetCalls();
-    execMock.mock.resetCalls();
-
-    // Mock getGitRoot and git commands
-    execMock.mock.mockImplementation((cmd: string) => {
-      if (cmd === "git rev-parse --show-toplevel") {
-        return Promise.resolve({ stdout: "/test/repo\n", stderr: "" });
-      }
-      if (cmd.includes("git -C") && cmd.includes("branch --show-current")) {
-        return Promise.resolve({ stdout: "\n", stderr: "" }); // Empty output = detached HEAD
-      }
-      if (cmd.includes("git -C") && cmd.includes("status --porcelain")) {
-        return Promise.resolve({ stdout: "", stderr: "" });
-      }
-      return Promise.resolve({ stdout: "", stderr: "" });
-    });
-
-    // Mock worktrees directory and contents
-    accessMock.mock.mockImplementation(() => Promise.resolve());
-    readdirMock.mock.mockImplementation(() =>
-      Promise.resolve(["detached-phantom"]),
-    );
-
-    const result = await listWorktrees("/test/repo");
-
-    strictEqual(result.success, true);
-    strictEqual(result.worktrees?.length, 1);
-    strictEqual(result.worktrees?.[0].name, "detached-phantom");
-    strictEqual(result.worktrees?.[0].branch, "(detached HEAD)");
-    strictEqual(result.worktrees?.[0].isClean, true);
-  });
-});
