--- src/commands/list.ts
@@ -1,1 +1,0 @@
-export { listWorktrees, type WorktreeInfo } from "../core/worktree/list.ts";
