--- src/commands/shell.ts
@@ -1,1 +1,0 @@
-export { shellInWorktree } from "../core/process/shell.ts";
