--- src/commands/where.ts
@@ -1,1 +1,0 @@
-export { whereWorktree } from "../core/worktree/where.ts";
