+++ src/core/git/libs/add-worktree.ts
