+++ src/core/git/libs/get-current-branch.ts
