+++ src/core/version.test.ts
